% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_0s_by_col.R
\name{rm_0s_by_col}
\alias{rm_0s_by_col}
\title{to remove all columns in matrix that have 0's in all rows}
\usage{
rm_0s_by_col(x)
}
\arguments{
\item{x}{Input vector, must be a matrix/data frame/array.}
}
\description{
to remove all columns in matrix that have 0's in all rows
}
\examples{
set.seed(123)
x = matrix(rpois(100, 0.1), 10, 10)
rm_0s_by_col(x)
}
