% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_0s_by_row.R
\name{rm_0s_by_row}
\alias{rm_0s_by_row}
\title{to remove all rows in matrix that have 0's in all columns}
\usage{
rm_0s_by_row(x)
}
\arguments{
\item{x}{Input vector, must be a matrix/data frame/array.}
}
\description{
to remove all rows in matrix that have 0's in all columns
}
\examples{
set.seed(123)
x = matrix(rpois(100, 0.1), 10, 10)
rm_0s_by_row(x)
}
